+++ src/elite/test/EliteSharpLib.Tests/SpaceTests.cs
@@ -405,7 +405,7 @@ public class SpaceTests
             out GameState gameState, out Universe universe, out _, out _, out _, out FakeEliteDraw draw, out RNG rng, out _);
         FakeShip oldPlanet = new(draw, rng) { Type = ShipType.Planet };
         universe.AddNewShip(oldPlanet, new(0, 0, 65536, 0), Matrix4x4.Identity, 0, 0);
-        gameState.Config.PlanetStyle = PlanetType.Wireframe;
+        gameState.Config.Game.PlanetStyle = PlanetType.Wireframe;
 
         space.RefreshPlanetStyle();
 
@@ -435,7 +435,7 @@ public class SpaceTests
             out GameState gameState, out Universe universe, out _, out _, out _, out FakeEliteDraw draw, out RNG rng, out _);
         FakeShip oldSun = new(draw, rng) { Type = ShipType.Sun };
         universe.AddNewShip(oldSun, new(0, 0, 64000, 0), Matrix4x4.Identity, 0, 0);
-        gameState.Config.SunStyle = SunType.Solid;
+        gameState.Config.Game.SunStyle = SunType.Solid;
 
         space.RefreshSunStyle();